#5568: use `BitInteger` to evaluate const int expressions with large types like u128

+++ src/main/kotlin/org/rust/lang/core/psi/ext/RsArrayType.kt
@@ -7,10 +7,11 @@ package org.rust.lang.core.psi.ext
 
 import org.rust.lang.core.psi.RsArrayType
 import org.rust.lang.core.stubs.RsArrayTypeStub
-import org.rust.lang.core.types.consts.asLong
+import org.rust.lang.core.types.consts.asInteger
 import org.rust.lang.core.types.ty.TyInteger
 import org.rust.lang.utils.evaluation.evaluate
+import java.math.BigInteger
 
 val RsArrayType.isSlice: Boolean get() = (greenStub as? RsArrayTypeStub)?.isSlice ?: (expr == null)
 
-val RsArrayType.arraySize: Long? get() = expr?.evaluate(TyInteger.USize)?.asLong()
+val RsArrayType.arraySize: BigInteger? get() = expr?.evaluate(TyInteger.USize)?.asInteger()
