+++ src/main/kotlin/org/rust/lang/core/types/ty/TyArray.kt
@@ -6,12 +6,13 @@
 package org.rust.lang.core.types.ty
 
 import org.rust.lang.core.types.consts.Const
-import org.rust.lang.core.types.consts.asLong
+import org.rust.lang.core.types.consts.asInteger
 import org.rust.lang.core.types.infer.TypeFolder
 import org.rust.lang.core.types.infer.TypeVisitor
+import java.math.BigInteger
 
 data class TyArray(val base: Ty, val const: Const) : Ty(base.flags or const.flags) {
-    val size: Long? get() = const.asLong()
+    val size: BigInteger? get() = const.asInteger()
 
     override fun superFoldWith(folder: TypeFolder): Ty =
         TyArray(base.foldWith(folder), const.foldWith(folder))
